+++ src/commands/ticker/index/processor.ts
@@ -12,7 +12,7 @@ import {
 import { InternalError } from "errors"
 import { Coin } from "types/defi"
 import { getChartColorConfig } from "ui/canvas/color"
-import { composeEmbedMessage } from "ui/discord/embed"
+import { composeEmbedMessage, justifyEmbedFields } from "ui/discord/embed"
 import {
   composeDaysSelectMenu,
   composeOtherTickerSelectMenu,
@@ -154,37 +154,47 @@ export async function renderSingle(
     ath,
     total_volume,
     max_supply,
+    total_supply,
   } = coin.market_data
   let icoDate = (coin as any)?.ico_data?.ico_start_date
-  const diff = moment.duration(
-    icoDate ? moment(moment.now()).diff(moment(icoDate)) : 0,
-  )
-  const age = icoDate
-    ? `${diff.years()}y${
-        diff.months() ? `${moment.duration(diff).months()}m` : ""
-      }`
-    : "N/A"
-  const fdv = max_supply
-    ? utils.formatUsdDigit(current_price[CURRENCY] * max_supply)
+
+  const maxSupply = max_supply || total_supply
+  const fdv = maxSupply
+    ? utils.formatUsdDigit((current_price?.[CURRENCY] ?? 0) * maxSupply)
     : "N/A"
 
   const current =
     type === ChartType.Dominance
       ? utils.formatPercentDigit(
           String(
-            (market_cap[CURRENCY] * 100) / total_market_cap[CURRENCY] ?? 0,
+            ((market_cap?.[CURRENCY] ?? 0) * 100) /
+              (total_market_cap?.[CURRENCY] ?? 0),
           ),
         )
       : utils.formatUsdPriceDigit({
-          value: current_price[CURRENCY] ?? 0,
+          value: current_price?.[CURRENCY] ?? 0,
           subscript: true,
         })
-  const marketCap = market_cap[CURRENCY] ?? 0
-  const embed = composeEmbedMessage(null, {
-    color: getChartColorConfig(coin.id).borderColor as HexColorString,
-    author: [coin.name, coin.image.small],
-    image: "attachment://chart.png",
-  }).addFields([
+
+  // dexscreener data
+  const { data: dexScreenerData }: any = await defi.searchDexScreenerPairs({
+    token_address: coin.contract_address,
+    symbol: coin.symbol,
+  })
+  const pair = dexScreenerData?.pairs?.[0]
+
+  // age
+  const diff = moment.duration(
+    moment(moment.now()).diff(moment(icoDate || pair?.created_at)),
+  )
+  const age = diff
+    ? `${diff.years() ? `${diff.years()}y` : ""}${
+        diff.months() ? `${moment.duration(diff).months()}m` : ""
+      }`
+    : "N/A"
+
+  const marketCap = market_cap?.[CURRENCY] ?? 0
+  const fields = [
     {
       name: `${getEmoji("CHART")} Market cap`,
       value: `${utils.formatUsdDigit(marketCap)} ${
@@ -204,24 +214,25 @@ export async function renderSingle(
       inline: true,
     },
     {
-      name: `${getEmoji("ANIMATED_FLASH")} Chain`,
-      value:
-        coin.asset_platform?.name ||
-        coin.asset_platform?.shortname ||
-        coin.name,
-      inline: true,
-    },
-    {
       name: `${getEmoji("ANIMATED_FLASH")} ATH`,
-      value: `${utils.formatUsdPriceDigit({
-        value: ath[CURRENCY] ?? 0,
-        subscript: true,
-      })}`,
+      value: ath
+        ? `${utils.formatUsdPriceDigit({
+            value: ath[CURRENCY] ?? 0,
+            subscript: true,
+          })}`
+        : "N/A",
       inline: true,
     },
     {
       name: `${getEmoji("ANIMATED_FLASH")} Volume (24h)`,
-      value: `${utils.formatUsdDigit(total_volume[CURRENCY])}`,
+      value: total_volume
+        ? `${utils.formatUsdDigit(total_volume[CURRENCY])}`
+        : "N/A",
+      inline: true,
+    },
+    {
+      name: `${getEmoji("ANIMATED_FLASH")} Max Supply`,
+      value: `${max_supply ? utils.formatDigit({ value: max_supply }) : "∞"}`,
       inline: true,
     },
     {
@@ -255,6 +266,15 @@ export async function renderSingle(
       value: age,
       inline: true,
     },
+    ...(pair
+      ? [
+          {
+            name: "Dex Screener",
+            value: `[${pair.name}](${pair.url.dexscreener})`,
+            inline: true,
+          },
+        ]
+      : []),
     ...(hasPlatforms && coin.asset_platform_id
       ? [
           {
@@ -264,7 +284,23 @@ export async function renderSingle(
           },
         ]
       : []),
-  ])
+  ].map((item) => ({
+    ...item,
+    value: item.value || "N/A",
+  }))
+  let embed = composeEmbedMessage(null, {
+    color: getChartColorConfig(coin.id).borderColor as HexColorString,
+    author: [
+      `${coin.name} (${
+        // coin.asset_platform
+        coin.asset_platform?.name || coin.asset_platform?.shortname || coin.name
+        // : ""
+      })`,
+      coin.image.small,
+    ],
+    image: "attachment://chart.png",
+  }).addFields(fields)
+  embed = justifyEmbedFields(embed, 3)
 
   const chart = await renderHistoricalMarketChart({
     coinId: coin.id,
@@ -392,7 +428,7 @@ export async function renderTokenInfo(
       {
         name: `${getEmoji("CHART")} Market cap`,
         value: `${utils.formatUsdDigit(
-          data.market_data.market_cap[CURRENCY],
+          data.market_data.market_cap?.[CURRENCY] ?? 0,
         )} ${
           data.market_data.market_cap_rank
             ? `(#${data.market_data.market_cap_rank})`